Update a comment
The old comment seemed to be inaccurate in multiple places. This
attempts to make it more accurate.

Of course, it's likely that the old comment was probably accurate at one
point and became inaccurate over time as the code changed. The same
could happen to this new comment. And since the inaccuracy of the old
comment wasn't noticed until now, it's arguable that we could remove the
comment altogether and not lose much.